+++ src/css/reviews.css
@@ -370,3 +370,53 @@
     font-size: 20px;
   }
 }
+.slick {
+  display: flex;
+  justify-content: center;
+}
+
+.slick-track {
+  display: flex;
+}
+
+.slick .slick-dots {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 13px;
+  margin-top: 32px;
+  margin-bottom: 32px;
+}
+
+.slick .slick-dots li button {
+  font-size: 0;
+  width: 13px;
+  height: 13px;
+  border-radius: 50%;
+  border: none;
+  background-color: rgba(17, 17, 17, 0.1);
+  transition: transform 250ms cubic-bezier(0.4, 0, 0.2, 1);
+}
+.slick .slick-dots li.slick-active button {
+  transform: scale(1.5);
+  width: 28px;
+  height: 13px;
+  border-radius: 30px;
+  background-color: orange;
+}
+
+.slick-slide {
+  margin: 0px 8px;
+}
+
+/* .slick-slide img {
+  width: 100%;
+} */
+
+.slick-active {
+  opacity: 1;
+}
+
+.slick-current {
+  opacity: 1;
+}
