+++ src/partials/reviews.html
@@ -18,17 +18,15 @@
       referrerpolicy="no-referrer"
     />
     <link rel="stylesheet" href="./css/reviews.css">
-    <link rel="stylesheet" type="text/css" href="slick/slick-theme.css"/>
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script type="module" src="./js/script.js"></script>
-    
+<link rel="stylesheet" type="text/css" href="https://cdn.jsdelivr.net/npm/slick-carousel@1.8.1/slick/slick-theme.css"/>
   </head>
  <section class="container-section" id="sellers">
       <div class="container">
         <h2 class="title">
           <span class="colored-title">CHOCOLATE</span> IS LOVED
         </h2>
-        <div class="carousel">
+        <div class="slick">
+        <div class="variable">
         <ul class="list">
           <li>
             <div class="container-cards">
@@ -148,30 +146,9 @@
           </li>
           </ul>
         </div>
-
-        <ul class ="list-two">
-          <li>
-            <svg width="28" height="13">
-              <use href="./images/icons.svg#Rectangle5"></use>
-            </svg>
-          </li>
-          <li>
-            <svg width="13" height="13">
-              <use href="./images/icons.svg#Ellipse3"></use>
-            </svg>
-          </li>
-          <li>
-            <svg width="13" height="13">
-              <use href="./images/icons.svg#Ellipse3"></use>
-            </svg>
-          </li>
-          <li>
-            <svg width="13" height="13">
-              <use href="./images/icons.svg#Ellipse3"></use>
-            </svg>
-          </li>
-        </ul>
+        </div>
       </div>
+</section>
     <div class="div-button">
         <button class="button" type="button" data-modal-open>Leave a review</button>
     </div>
@@ -231,8 +208,23 @@
         </form>
       </div>
     </div>
-    <script type="module" src="https://ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
-    <script type="module" src="./slick/slick.min.js"></script>
+    </html>
+    <script
+      src="https://code.jquery.com/jquery-2.2.0.min.js"
+      type="text/javascript"
+    ></script>
+    <script src="https://cdn.jsdelivr.net/npm/slick-carousel@1.8.1/slick/slick.min.js"></script>
     <script type="module" src="./js/script.js"></script>
-</section>
+    <script src="./js/slick.js" type="module" charset="utf-8"></script>
+    <script type="text/javascript">
+      $(document).on('ready', function () {
+        $('.variable').slick({
+          dots: true,
+          infinite: true,
+          variableWidth: true,
+          arrows: false,
+          dotsClass: 'slick-dots',
+        });
+      });
+    </script>
 </html>
